requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.1.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
cron-descriptor==1.4.5
cryptography==42.0.8
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
Django==5.0.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.11.2
django-extensions==3.2.3
django-nine==0.2.7
django-rest-knox==4.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
executing==2.0.1
gunicorn==23.0.0
h11==0.14.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
isort==5.13.2
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.4.0
lorem==0.1.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.1
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
pydevd-pycharm==242.19890.32
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.1.1
pyOpenSSL==24.2.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
redis==5.0.8
requests==2.31.0
selenium==4.20.0
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tomlkit==0.12.5
tornado==6.4.1
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
